requirements.txt
# Core Testing Framework
pytest==7.4.3
pytest-cov==4.1.0
pytest-html==4.1.1
pytest-xdist==3.3.1
pytest-mock==3.12.0
pytest-asyncio==0.21.1

# Data Quality and Validation
great-expectations==0.17.23
pandas==2.1.4
numpy==1.24.3
pydantic==2.5.0
jsonschema==4.20.0

# Database Testing
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymysql==1.1.0
sqlite3

# API Testing
requests==2.31.0
httpx==0.25.2
fastapi==0.104.1
uvicorn==0.24.0
pytest-fastapi==0.0.1

# Performance Testing
locust==2.17.0
pytest-benchmark==4.0.0
memory-profiler==0.61.0

# Data Processing
apache-airflow==2.7.3
apache-kafka-python==2.0.2
pyspark==3.5.0
dask==2023.11.0

# Monitoring and Alerting
prometheus-client==0.19.0
grafana-api==1.0.3
sentry-sdk==1.38.0

# Configuration Management
pyyaml==6.0.1
python-dotenv==1.0.0
configparser==6.0.0

# Data Analysis and Reporting
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
dash==2.14.2

# Utilities
click==8.1.7
rich==13.7.0
tqdm==4.66.1
python-dateutil==2.8.2
pytz==2023.3

# Security Testing
cryptography==41.0.8
bcrypt==4.1.2

# CI/CD Integration
jenkinsapi==0.3.13
github3.py==4.0.1

# Containerization
docker==6.1.3

# Logging
structlog==23.2.0
loguru==0.7.2

# Async Support
aiohttp==3.9.1
asyncio-mqtt==0.16.1

# Machine Learning (for anomaly detection)
scikit-learn==1.3.2
scipy==1.11.4

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0 